+++ test/phoenix/controller/render_test.exs
@@ -23,7 +23,7 @@ defmodule Phoenix.Controller.RenderTest do
     assert conn.resp_body == "Hello\n"
     assert html_response?(conn)
     refute conn.halted
-    assert controller_template(conn) == "index.html"
+    assert view_template(conn) == "index.html"
   end
 
   test "renders atom template" do
@@ -32,7 +32,7 @@ defmodule Phoenix.Controller.RenderTest do
     assert conn.resp_body == "Hello\n"
     assert html_response?(conn)
     refute conn.halted
-    assert controller_template(conn) == "index.html"
+    assert view_template(conn) == "index.html"
   end
 
   test "renders string template with put layout" do